fix: Fix id string

+++ backend/school-management/app/Core/Application/UseCases/Payments/Student/Cards/DeletePaymentMethoUseCase.php
@@ -5,6 +5,8 @@ namespace App\Core\Application\UseCases\Payments\Student\Cards;
 use App\Core\Domain\Repositories\Command\Payments\PaymentMethodRepInterface;
 use App\Core\Domain\Repositories\Query\Payments\PaymentMethodQueryRepInterface;
 use App\Core\Domain\Repositories\Stripe\StripeGatewayInterface;
+use App\Exceptions\NotFound\PaymentMethodNotFoundException;
+use Illuminate\Support\Facades\DB;
 
 class DeletePaymentMethoUseCase
 {
@@ -19,8 +21,17 @@ class DeletePaymentMethoUseCase
     public function execute(int $paymentMethodId): bool
     {
         $paymentMethod=$this->pmqRepo->findById($paymentMethodId);
-        $this->stripe->deletePaymentMethod($paymentMethod->stripe_payment_method_id);
-        $this->pmRepo->delete($paymentMethod->id);
+        if (! $paymentMethod) {
+            throw new PaymentMethodNotFoundException();
+        }
+        DB::transaction(function () use ($paymentMethod) {
+            $this->stripe->deletePaymentMethod(
+                $paymentMethod->stripe_payment_method_id
+            );
+
+            $this->pmRepo->delete($paymentMethod->id);
+        });
+
         return true;
     }
 }
